Spur: Reimplement deriving the accessorDepth and retrying primitives on primitive failure machinery.  Always take the primitive index from newMethod (setting newMethod to a SmallInteger for primitiveDoPrimitiveWithArgs.
Have the Cogit always set primitiveFunctionPointer and
newMethod for the retry,

Add isNonInteger(Non)Immediate: and use it in several of
the integer conversion routines that were missing code
to exclude Character immediates, with fatal consequences.

With these changes the Pharo 4 test suite runs without
crashing on the latest Pharo Spur boot image.