VMMaker.oscog-eem.924 Simulator/Slang: Refactor to allow 32-bit and 64-bit Spur to coexist in the simulator, thereby allowing a 64-bit bootstrap.
Replace direct use of BytesPerWord, ShiftForWord,
BytesPerOop, and BaseHeaderSize with sends of
wordSize, shiftForWord, bytesPerOop and
baseHeaderSize respectively.
Make sure these are still output as the manifest
constants in the generated C.  Eliminate use of
bytesPerSlot in favour of bytesPerOop.  The use of
sends is similar to those in trunk VMMaker, but I
don't have time to merge.  Sorry :-(.

Work-around use of sends for bytesPerOop in e.g.
generateToByDo:on:indent: and users of
isConstantNode:valueInto:.

Do a better job at inlining via
inlineCodeOrNilForStatement:in:.  Comment in
isNode:substitutableFor:inMethod:in: explains.

Eliminate generic VM_LABEL support, labelling only
the bytecodes in interpret.

Eliminate use of LargeContextSize in the stack and
cog VMs.  Alas haven't yet eliminated the frame
offset constants such as FoxCallerSavedIP et al.
We only need one of the two word sizes of
interpreter to be executable in the bootstrap.