VMMaker.oscog-eem.966
Add a -breakmnu vm argument that calls warning on an mnu of a specific selector.

Fix SmartSyntaxPlugin code generation for Spur of

- asOop: old code assumed 4 byte oops & headers
- Unsigned old code assumed 32-bit oops


Affects AsyncFilePlugin, BochsIA32Plugin & FT2Plgin.

Avoid duplication of effort in special selector send forwarding fault handling.

Improve stack trace printing for Newspeak and Glue by providing something that
assumes the penultimate literal is the selector or an AdditionalMethodState-like
thing containing it.

Reimplement primitiveCopyObject to work for both pointer and bits objects.
This to support a good Object>>clone for Spur.

Fix Spur's isWordsOrBytes[NonImm]:.

More changes to the Mac platform subsystem for 64-bits.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3159 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 44f1c3e5299fb5b817155fedc1c72ba2905ccdd0