CogVM source as per VMMaker.oscog-eem.972
Fix a devilishly tricky bug with 32-bit Spur and PICs in the Cogit.  In 32-bit
Spur only zero is a valid cache tag, that for Characters, due to SmallIntegers
having both 2r11 and 2r01 as valid tags.  The cache check prolog collapses
these down onto 1 by anding with 1 and hence collapses Character's 2r10 to 0.

The cache tag is also tested in the PIC abort routine to distinguish between
an MNU abort and an interpret abort.  A PIC MNU entry jumps to code that first
zeros the cache tag (held in ClassReg) before calling the abort routine, which
tests ClassReg and selects the relevant abort processing.

Hence in 32-bit Spur if a send site is linked for Character and later extended
to a PIC with an interpret case, that interpret case will instead invoke MNU
processing.  This is rare; first there needs to be a send to a CHaracter, then
a send to something else which binds to something that must be interpreted.
Ouch.

The fix is simple; if a send site that was linked with a zero cache tag (which
can only happen in 32-bit Spur) link to an open PIC instead of to a closed PIC,
or in the case of an MNU do not link.

Changes:
Rename compilePICProlog:, mnuCall and interpretCall to the more accurate
compilePICAbort:, picMNUAbort and picInterpretAbort.  Coment compilePICAbort:
and compileAbort. Name the zero value picAbortDiscriminatorValue to make the
code more comprehensible/browsable.

Assert that the cache tag is not zero in ceCPICMiss:...

Refactor cogMNUPICSelector:methodOperand:numArgs: to add receiver: parameter
for the check for a zero cache tag.

Eliminate some duplication introduced in early versions of the ARM abort code.
Specifically lose sendMissCall mnuCall interpretCall & interpretLabel.

Fix type regression in cogMethodOf: with cCoerceSimple:to:; CogClass's version
should be marked <doNotGenerate>.

Fix type warnings in primitiveCopyObject and printStackCallStackOf:.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3163 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 62a91034a579034d2f9ca4bd53e746c63634f3e7