Use the gcc/clang built-ins for CAS and atomic increment if available.
A few more 64-bit changes to the Carbon implementation of the Mac subsystem,
taking it as far as I can before admitting defeat and switching to the Cocoa
implementation.

Add a linux 64-bit build tree.

Shed the old external platforms/iOS tree and import the latest pharovm tree
(as of:
	commit 4feca169ed2abe8b3df0c0d0b9cd1f980a84e723
	Merge: 447f247 3c7a588
	Author: Esteban Lorenzano <estebanlm@gmail.com>
	Date:   Fri Dec 5 07:54:51 2014 +0100

		Merge pull request #70 from nicolaihess/issue_14340
		
		use the same value for charCode and keyValue for mousewheelevents)



git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3166 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: b5f20d94a165702b033841443716fa12de18393b