CogVM source as per VMMaker.oscog-eem.981
CHanges for 64-bit compilation.

Eliminating warnings in B2DPlugin and the VMProfileLinuxSupportPlugin.

Handle not in nilOrBooleanConstantReceiverOf:.  Hence don't do the 32-bit swap
in a 32-bit words object in 64-bit Spur.  May have to do it in a 64-bit word
obj in 64-bit Spur, but it can wait.

Generate the right casts for left shift beyond the int range.  Generate the
right error code for negative arguments to primitiveSizeInBytesOfInstance

Correctly translate constant arithmetic expressions that overflow the int range.

The criterion for promoting a shifted constant to long was wrong.  It needs to
compare against sizeof(int) (the default type) not sizeof(long).


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3177 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: cf1b6e4df263eee21d8cbd386388bcf82435d748