VMMaker.oscog-eem.1004.mcz Revisit the bug fix for followed contexts during scavenging in VMMaker.oscog-eem.913 of 24 October 2014.  Split isWidowedContext: into a more careful version, isWidowedContextDuringGC: for use only during scavenging.
Speed up non-local return processing by using the
fact that 98% of the time returns are within the
same page.  The search for the home context
following finding an unwind is unnecessary iff the
home is married and on the same page.

Refactor commonReturn to eiminate CoInterpreter's
version, moving the divergence into
maybeReturnToMachineCodeFrame.

Makew sure the search for the home through the
block's closure uses followField:ofObject:.

Fix a slip in genBinaryInlineComparison:opFalse:destReg:.

Nuke erroneous instance-side versions of
initializeFrameIndices (editing blunders).

Typos.