Spur: Make sure the saved contexts at the top of each stack page are followed post-become in followForwardingPointersInStackZone:.
Add an assert to Spur's shouldRemapObj: to verify
that new space references are only visited once in
a scavenge.

Add asserts to check that contexts are not
forwarded when marrying stable contexts.

General:
Never inline mapStackPages, for debugging.

Add an assert to returnToExecutive:postContextSwitch:
to ensure the instructionPointer is valid when
checking for being at a send return pc.

Eliminate the simulator overrides for
withSmallIntegerTags: and don't assume the
SmallInteger tag pattern.