CogVM source as per VMMaker.oscog-eem.1048
Refactor leak checking code so that GCModes are flags and coincide with flags
for what GC events to leak check.  Hence ensure leak checking is actually run
when requested.  As part of this move the shrink action from postGCAction:
into the memory managers and hence allow Spur to not shrink on scavenge.

Fix bugs with immediate character printing.
Allow printing the mark state of free chunks.

Spur:
Fix some dumb slips in the image segment code.

Only copy-and-forward in remapObj: in the right cicumstance.
Add an assert to remapObj:.

Have lengthOf: answer 0 for ephemerons and
something sensible for 64-bit indexable on 32-bits.

Scavenging ephemerons should update the ephemeron's key with the scavenged key.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3249 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: c38721d78db6ef1ab283c6d32f60be1c6d8937aa