VMMaker.oscog-eem.1071.mcz Fix several bugs in the new type inferrence code, resulting in a functional VM again (phew!).
Do a much better job at typeForArithmetic:in:;
answering the longest type for integral types.

In inferTypesForImplicitlyTypedVariablesIn:, observe
assignments from both sends and typed vars.
Comment the method better so that 
alreadyExplicitlyTyped's usage is clear(er).

On type compatibility of inlined expressions,
insist on signed compatibility or accept an untyped
actual.

Provide a return type for asFloat (so inlining in
primitiveAsFloat still occurs)

Provide an exception in unsignedTypeForBitwiseSend:in:
for currentBytecode to preserve the nice constant
inlining of currentBytecode & N in the bytecode cases.

This time I feel that the Slang Sensei has hit me
over the head and some enlightenment has ensued.
That's a first ;-).