CogVM source as per VMMaker.oscog-eem.1072
Fix typing multiple assignments to a variable
- promote the type to the largest integral type of all assignments
	(primitiveDecompressFromByteArray)
- do /not/ promote a variable typed with integral type to a floating point type
	(transformColor:)

Provide size of short int.

Do a much better job at typeForArithmetic:in:;
answering the longest type for integral types.

In inferTypesForImplicitlyTypedVariablesIn:, observe assignments from both sends
and typed vars.  Comment the method better so that alreadyExplicitlyTyped's
usage is clear(er).

On type compatibility of inlined expressions, insist on signed compatibility or
accept an untyped actual.

Provide a return type for asFloat (so inlining in primitiveAsFloat still occurs)

Provide an exception in unsignedTypeForBitwiseSend:in:
for currentBytecode to preserve the nice constant
inlining of currentBytecode & N in the bytecode cases.

Improve the type inferrence machinery further to
support pointer subtraction.  Eliminate code duplication
(determineTypeFor:in: now uses returnTypeForSend:, etc).

Eliminate the funky code generation scheme for translated primitive plugins
(MiscPrimitives et al) and use the same inlining and code generation path as
normal plugins.  Finally implement a time stamp checker that looks at the
source method timestamps.

Type some InterpreterProxy methods to generate valid signatures.

...and add type inference support for at: to rescue
inlining in the Klatt plugin.

Add support for type inferrence of simple arithmetic
(+,-,*,/) but *not* >> << et al, so that inlining is still
done for e.g. pointer arithmetic expressions.

Fix a bug in TMethod>>typeFor:in: which was causing
the failure to find the types of global variables.

Change the heuristic for inlining expressions to include a maximum usage count.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3260 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 37986cd5b49c236a1f6f5819fe085df3b0a9c5c1