VMMaker.oscog-eem.1086.mcz Simulation: Fix regression in simulator's isScreenSize.
Implement simulation of translated primitives that
don't access inst vars (the bitmap (de)compression
and string primitives).  So that the machinery can
match actual selectors with primitive selectors
have a TMethod remember the original selector of
a translated primitive TMethod in properties.

Add an assert to InterpreterStackPages>>longAt:put:
to prevent storing garbage there-in when there's
a bug.