CogVM source as per VMMaker.oscog-eem.1111
Fix misgeneration of saveAndRestoreLinkRegAround: by not marking as required
<inline: true> methods in CogAbstractInstruction.  Has the nice side-effect of
deleting ~ 2000 lines of dead code from cogit.c.  Revert the bogus fix to
preGenerationHook:.

Fix type coercion in propagateReturnIn: to not generate unnecessary casts.

Do dead-code elimination on (expr and: [false]) ifTrue:
and hence fix a few exprs in primitiveVMMarameter

Merge in Nicolas' fixes for incorrect use of #|.

Refactor the enilopmart creation machinery so that there is only one creation
routine, adding a forCall: parameter to differentiate enter and call
enilopmarts.

Fix CogMethodZone>>whereIsMaybeCodeThing: which was testing the wrong boundary.



git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3283 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 3618512396ea5ba7fa76bbb6fdbdabd4aa47354c