CogVM source as per VMMaker.oscog-eem.1118
Work around a compiler bug in 64-bit gcc 4.4.7-4 on RedHat that elided one of
the two assignments to a corpse's header in threading it onto a weak survivor
list.  Do so by, in 64-bits, performing a single assignment to the corpse's
header, instead of two.

Mark copyToOldSpace:bytes:format: as not to be inlined so as to avoid impacting
copyAndForward's icache density.

Simplify the macros in sqMemoryAccess.h to avoid the double cast on the
oop argument of fooAtPointer[Put].


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3288 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: f81d10e82e83ec83d323a594b45224a50d4fa6eb