VMMaker.oscog-eem.1127 Rename executeCogMethod:... to executeCogPIC:...
Add support for a variable base register for
referring to the CoInterpreter variables from
machine code on RISCs (ARM).  Load the VarBaseReg
in enilopmarts.  It must be a callee-saved reg.

Add support for a 3 argument AndCqRR and use it
in the entry-point sequence.  This has the nice side-
effect of fixing the inability to differentiate the
checked and unchecked entry-points on ARM if we
use the MoveRR, AndCqR two instruction sequence ;)

Nuke the unused and misleading callsAreRelative.

Mark the other backEnd ability test methods as
<inline: true>.

Simulator:
Add print cog methods with selector... to the utils menu.