CogVM source as per VMMaker.oscog-eem.1155
and fix that compiler warning properly.  If one defines
a function return type the arguments are typed there-in.

Sista:
Fix inline primitive for at: and at:put: to be 1 relative. This simplifies the
JIT in 64 bits because the 1 relative correponds to the 1 word difference
between baseHeader and first object field hence generating one less instruction.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3305 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 4f82106df5aeefb8764803cc7cee5c9e96937b35