WARNING!! Because of the rename of VMSqueakV3BytecodeConstants to VMBytecodeConstants you will probably have to do a recompile package after loading this iunless loading it for the first time.
Add support for a directed super send bytecode
that starts the lookup above the class that is the
value of an association pushed on the stack immediately
before the send.  This bytecode is used in SistaV1
to allow inlining of code that contains super sends.

Add mapFor:bcpc:withAnnotationPerformUntil:arg: and
use it to gather send and branch data in the Sista
Cogit.

Rename VMSqueakV3BytecodeConstants to
VMBytecodeConstants and add the
BytecodeSetHasDirectedSuperSend flag.  Rewrite
initialization of NumTrampolines to take into account
the directedSuperSend trampolines.

Use PushCq to pass small integer constants in the
trampolines instead of smashing TempReg, which is
now used to hold the directed supersend argument.
Add support for PushCq.  Could use PushCw but on
x86 PushCq is a lot shorter.

Change a few byteAt: sends to send to the right class.
Fix a slip in isDirectedSuper:extA:extB:.  Fix slip in
annotationForSendTable:.
Fix a bug in nextDescriptorAndExtensionsInto: so that it
answers info for the next, not the current, bytecode.
Fix printing in CurrentImageCoInterpreterFacade so that
printTrampolineTable works with a facade.