Fix code generation bug surfaced by inline primitives. On x86 movb N(%reg),%rl can only store into al, bl, cl & dl, whereas movzbl can store into any reg.  On ARM move byte also zero-extends.  So change definition of MoveMbrR to always zero-extend, use movzbl on x86 and remove all the MoveCq: 0 R: used to zero the bits of the target of a MoveMb:r:R:. And now that we have genGetNumSlotsOf:into:, use it.
Fix CogIA32CompilerTests for testCmpRdRd and
modify testMoveMbrR appropriately.

Fix a slip in genTrinaryInlinePrimitive:, meet the constraint
that the target must be in ReceiverResultReg, and do
a better job of register allocation there-in.

Do dead code elimination for the branch following
an inlined comparison (this is done in
genBinaryInlineComparison:opFalse:destReg:
copying the scheme in genSpecialSelectorEqualsEquals).