CogARMCompiler: Fix updating of call and jump instructions by splitting the call/jump update from the callfull/jumpfull update.  Rename the operand extractor to extract32BitOperandFrom4InstructionsPreceeding: and add insert32BitOperand:into4InstructionsPreceeding: to go with. Use these to update instructions.
Modify machineCode array to be array of longs. Refactor
machine code output so that CogARMCompiler can spit
out words while the x86 dribbles bytes.

Avoid annotating the absolute CallFull/JumpFull instructions
so that the update is simplified.

Now the system survives a code compaction on ARM
although it doesn't continue for long there-after.