In order to fix the SistaCogit (still crashing somehow), I refactored the #== code which was duplicated three times between the Stack v3 / Stack Spur / Sista Spur cogits, mainly by creating 2 smaller methods which are now called by the 3 duplicated methods.
I introduced: #extractMaybeBranchDescriptorInto: so I don't need to duplicate the code that fetch the branch descriptor in #== and comparison.

Removed a useless allocateCallReg in genStorePop:RemoteTemp:At: because the call reg are saved in the trampoline.

Removed a duplicated self assert: needsFrame in genStorePop: popBoolean LiteralVariable: litVarIndex (Eliot moved it to the beginning of the method).