VMMaker.oscog-cb.1257 Extracted the count logic in SistaCogit to be in one method called by the 3 methods generating branches to avoid duplication.
I believe there's a bug in JumpIf:, if the couterReg is callerSaved then I think it is not saved correctly. That was probably not an issue before I introduced the better register allocation because Eliot most probably carefully picked a register for counterReg which was callee saved in x86.

Added a few comments and indentation fix in StackToRegisterMappingCogit.