CogVM source as per VMMaker.oscog-eem.1292
Fix a regression in externalInstVar:ofContext:.  The head frame pointers must
be written back if we're going to map a machince code pc to a bytecode pc in
case of code reclamation.

Add asserts to the stack page enumerators to check
that the head frame pointers have been written back.

Use macros for the oop comparisons, avoiding cCoerce:, to get faster simulation
and avoid the inliner not always inlining in conditionals.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3322 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 2969a404cc95f787c2d8251bfd2c1ae472200817