VMMaker.oscog-eem.1309.mcz Fix ceActivateFailingPrimitiveMethod: to return properly after retrying a primitive after checkForAndFollowForwardedPrimitiveState. e.g. fixes | s | 1 to: 5 do: 	[:i| 	s := 'bar'. 	s becomeForward: 'bzzt' copy. 	'foo' <= s]
Fix bug in followForwardedFrameContents:stackPointer:
so that arguments are actually followed.  N.B. drives
the above bug underground ;-).

Make the bounds check in primitiveObject:perform:...
more restrictive and fix bugs in the commentary.

Make stack pages a reasonable size.  My math was
wrong and the size was half as big as intended.

Simulator:
Refactor accessor depth determination of accessors
so that simulator derives correct values for translated
primitives.  And fix
translatedPrimitiveArgument:ofType:using: so that
translated primitives will fail in the simulator.