CogVM source as per VMMaker.oscog-eem.1315
Slang:
Do a more thorough job of transforming complex to:[by:]do: loops so that the
limit expression is only evaluated once, not on each iteration of the loop.

Fix regression in read-before-written code when
rewriting to only initialize simple types.

Cogit:
Fix ARM-only regression in StackToRegisterMappingCogit's
PIC aborts when the SimpleStackBasedCogit was fixed.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3350 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 932e583421ca34bfd31cfeb7f880b694a772f3f5