CogVM source as per VMMaker.oscog-eem.1333
Cogit:
Fix assert-fails with absent receiver sends in Newspeak.  Have the implicit
and outer send lookup trampolines set the stacked receiver (when there is one)
when setting the implciit receiver.

Streamline send trampoline creation by refactoring trampoline name generation
so that the two limits, NumSendTrampolines - 2 and numRegArgs are treated
separately, and numArgsOrSendNumArgsReg: answers the relevant numArgs argument.
This should clear up confusion between numRegArgs (which can be 0, 1 & 2) and
NumSendTrampolines - 2, which is always 2.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3367 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 3e7dc7f3e111dc9adc0d9c6aa46468ad271c8ca2