VMMaker.oscog-eem.1344.mcz Cogit: Get code generation working for CogOutOfLineLiteralsARMCompiler.
Fix initialization of methodLabel in 
Introduce DumpJumpLong[Zero]: for convenient
PIC wirting where the literals need to be dumped
in consistent places to keep PIC cases the same size.

Fix isAnInstruction for labels (methodLabel) and
use it to guard addressIsInCurrentCompilation:
calls where a label is involved.

For better initialization of literal management
use zeroOpcodeIndex everywhere opcodeIndex
is initialized, and have it reset literals.  And have
computeMaximumSizes dump literals.

Move the assignment of maxSize out of
computeMaximumSize and into
computeMaximumSizes for concision.

ARM Cogit:
Introduce some more abstractions for rotateable
constants so that it is easier to relate
computeMaximumSize and dispatchConcetize in
CogARMCompiler.  Provide more opcode names and
use them, and use inverseOpcodeFor: to calculate
the inverse opcode when the inverse constant is
used.

Revise ARM getDefaultCogCodeSize given better
estimates of relative code sizes.
Simplify isPICDependent to eliminate Calls, which
are not pc-dependent (ince they always call out of
the current compilaiton).

Simulator:
Safety in instruction printing, and spotting literals
in disassembly.

Slang:
Fix ancilliaryClasses: given the deeper
CogAbstractInstruction hierarchy on ARM.