CogVm source as per VMMaker.oscog-eem.1349
Slang:
Modify prototype output so that non-static functions are declared as extern.
This to try and cure Cog crashes on compilers such as clang 3.4.

Spur:
Remember to count shrink requests

Cogit:
Changes for in-progress CogOutOfLineLiteralsARMCompiler code generator.

Move annotations into instructions, where they sit more naturally anyway.
But doing so makes it trivial to move annotations from the instructions
referencing out-of-line literals to the literals themselves.  Hence nuke
CogInstructionAnnotation.

Fix isAnInstruction for labels (methodLabel) and use it to guard
addressIsInCurrentCompilation: calls where a label is involved.

Move the assignment of maxSize out of computeMaximumSize and into
computeMaximumSizes for concision.

ARM Cogit:
Introduce some more abstractions for rotateable constants so that it is easier
to relate computeMaximumSize and dispatchConcetize in CogARMCompiler.  Provide
more opcode names and use them, and use inverseOpcodeFor: to calculate the
inverse opcode when the inverse constant is used.

Revise ARM getDefaultCogCodeSize given better estimates of relative code sizes.
Simplify isPICDependent to eliminate Calls, which are not pc-dependent (since
they always call out of the current compilaiton).

Fix bugs in computeMaximumSize with AddCqR and
SubCqR, and a bad slip in concretizePushCq.

Nuke the unused BICCqR opcode.  Add class vars to
name all the opcodes passed to concretizeDataOperationCwR:.

Make all the usages of flushICacheFrom:to: work with start-address to
byte-after-end-address. Irrelevant to x86, perfect for ARM, matches most of
the usage already in place.

Sista Cogit:
Added support for jumpOrPopIfNotInstanceOf: in cogit's scanner.

Unix: restore the SCCS revision after the version number in -version output.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3374 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 9b98735ac8bc04f499f162a7a36bdb3755f30f1e