Update sources for VMConstruction-Plugins-OSProcessPlugin-dtl.40
OSProcessPlugin 4.6.1

Critical bug fix by Thierry Goubier: Use of vfork() in #primitiveForkExec leads to intermittent failures because child is modifying memory space prior to calling an exec. Use fork() instead because it protects against this, and because any performance benefit of vfork() is no longer relevant. Also add error check for the fork() call to fail the primitive properly.


git-svn-id: http://squeakvm.org/svn/squeak/trunk@3375 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 5c096e6b30727bd4ecd5b604dd06ffb890853e0f