CogVM source as per VMMaker.oscog-eem.1354
Cogit:
Fix bug in rewriting compiler primitives on module unload, etc.  The old code
didn't change the assignment to the primitiveFunctionPointer, which is needed
for correct management of failing primitive calls on Spur.  Simplify the post
compile hook to eliminate the label parameter; this is local to the Cogit.

ARM Cogit:
Replace CogInLineLiteralsARMCompiler with CogOutOfLineLiteralsARMCompiler.

Fix some bugs with PushCw, and AndCqR.

Move the calculation of which address to annotate for an annotated instruction
into the instruction, allowing CogOutOfLineLiteralsARMCompiler to annotate the
out-of-line literal, rather than the instruction following, and implement
literal update on GC for out-of-line literals.

Slang:
Eliminate blank lines coming from conditional dead code elimination.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3376 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 6320baa287e8ee0eaf286983acc5041f9a985b74