CogVM source as per VMMaker.oscog-eem.1401
Spur:
Provide a limit to the size of old space on Spur.  Access maxOldSpaceSize
via vmParameterAt: 67 & command line parameter -maxoldspace.

Cogit:
Simplify pushRemoteTemp register management to be able to improve further the
overall register management.

Have optStatus hold onto the sim stack entry it is holding, if any.

Sista Cogit:
fixed trinaryInlined primitive which was forcing a reg to be ReceiverResult reg when it shouldn't.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3393 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 45b64d4c6f112b413d13fb057bd29540d6a2ba31