CogVM source as per VMMaker.oscog-eem.1408
ThreadedFFIPlugin:
Allow ffiCreateIntegralResultOop:ofAtomicType:in: to be inlined.  Reduce the
number of tests leading to the common case of an integral return in
ffiCalloutTo:SpecOnStack:in: etc.

Refactor the remapOop:in: idiom up into InterpreterPlugin and use it in the
ThreadedFFIPlugin.

Use an official version of MiscPrimitivePlugin.c in Newspeak (no change).


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3400 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: a64faca47a8af0d3224f4c41be51460a99864470