CogVM source as per VMMaker.oscog-eem.1449
Fix positive32BitValueOf: & signed32BitValueOf: for 64-bit Spur;
these must fail for SmallIntegers with digitLength > 4.

Modify secret primtiive 161, primitiveSetIdentityHash so that with 0 args it
answers whether an object has an identity hash.  This is vacuously true for V3,
but is meaningful in Spur.

Update the Newspeak FilePlugin to be the same as the Smalltalk one.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3434 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: c30e8098e96154cdcec2f00700432b23c9e23afd