VMMaker.oscog-eem.1476.mcz x64 Cogit: Implement most of the inline cache tag/literal extraction code so that CogInLineLiteralsX64Compiler can generate both PIC prototypes.  This is broken because I was confused about PushCw, thinking there was a push 64-bit immediate, there isn't.  I'll have to undo some.
Change the 64-bit surrogates to use long64At:[put:] instead of the deprecated unsignedLongLongAt:[put:].

Fix a bad (but currently symptomless) bug in classRefInClosedPICAt: which had things backwards.

In-line compilation now "works" for CogInLineLiteralsX64Compiler (pace the PushCw confusion), except for the use of 32-bit entries in in-line caches.  hence the next task is to compile sends so that 32-bit inline caches can refernece selectors (i.e. by containing indices into a method's literal frame or the special selector array).