VMMaker.oscog-eem.1481.mcz Slang: Propagate types of actuals with other than the default sqInt type to untyped formals, hence fixing bogus inlining of genLoadHeaderIntoNewInstance: into genGetActiveContextLarge:inBlock:.
Fix baaaad bug in node:typeCompatibleWith:inliningInto:in: which was looking up the type of the formal parameter in inlining in the target method, not the method being inlined! (bug surfaced by previous fix).

Allow constant elimination for Const ifNil: [] ifNotNil: [], hence cleaning up uses of characterTable in the Spur VMs.