VMMaker.oscog-eem.1522 Cogit: Implement the long conditional branch/long unconditional branch split necessitated by the MIPS processor in all of the closed PIC methods.
Update Slang to collapse an ifTrue:ifFalse: if bioth arms are the same, to avoid the code duplication this introduces on the rest of the processors where conditional and unconditional branch offsets can be accessed in the same way.

Caution:  Tim's new CPICs are broken w.r.t. accessing class tags in PICs.  In fact, he's left the breaks in classRefInClosedPICAt: & storeClassRef:inClosedPICAt: to show his unease.  I'll fix this asap.  First I need to add code to disassemble the closed PIC prototype so I can see the wood for the trees.