CogVM source as per VMMaker.oscog-eem.1539
Cogit:
Fix /horrible/ bug with primitive error codes and fixups.  The adjustment of the
initialPC to skip the primitive and error code, if any, was done after
generating fixups for backward branches, and hence these branches were bogus.
Object>>shallowCopy in Spur is subject to this, but amazingly enough it has not
surfaced before now.  Fix this by adjusting the initialPC before scanMethod
scans for fixups.

ARM builds:
Revert inclusion of Squeak3D plugin until the code is in the src/plugins tree.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3511 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: e96d89d6815913f4bb5302dfa58b6d783f45552d