CogVM source as per VMMaker.oscog-eem.1558
Cogit:
Refactor the logic to handle UnimplementedPrimitive and UnfailingPrimitive in a
common routine, passing the result of the primtiiev generator to
#compileFallbackToInterpreterPrimitive:.  Add the notion of CompletePrimitive
for primitives that may fail, but handle all the success cases, hence if a
method including such a primitive doesn't use the error code the Cogit doesn't
have to compile the call to the interpreter primitive as a fallback.

Add genPrimitiveObjectAt to provide fast literal and header access on Spur.

Streamline access to the tag patterns form the Cogit by using macros.

CPIC relocation: fix conflation of unconditional and conditional jumps.

Implement overflow checking for primitive div that works in 32- and 64-bits.

Provide a convenience method to hide the inconvenience of checking for
UndefinedOperation in uses of genJump[Not]SmallInteger:.

Refactor to share genInnerPrimitiveStringAt: between 32- and 64-bit Spur obj
reps.

x64 Cogit: rationalise concretize"Op"RR into two routines in both x64 and x86.

Avoid using isCallPrecedingReturnPC: for testing if compileInterpreterPrimitive
used a CallFull or a JumpFull to invoke the prim, hence allowing
isCallPrecedingReturnPC:to only have to apply to calls used for sends, and
hence reuse x86's isCallPrecedingReturnPC: for x64.

Fix genPrimitiveClosureValue for 64-bits (use genJumpImmediate: to exclude
non-cogged methods).

Plugins: Include the file sync prim in the FilePrims plugin.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3514 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 83855df1a2104838877f765dca263afd9d15476f