Initial commit where I added immutability in the VW's sense to the Cog VM.
Most code is under the IMMUTABILITY compilation flag and does not modify the regular VM. We note the exception of some machine code store generation that were simplified to support immutability. I checked the Spur and non Spur Cog VM should work fine.

The immutability callback is expected to return without pushing any value after the store (the callback needs to manually do the failed store if needed). This was more convenient in the case of popInto as execution resumes after the popInto instruction.

TODO:
- make all the needed store instructions mapped (isMapped) in Cogit. Maybe a better way than what I started using isMappedIfImmutable as Eliot suggested.