VMMaker.oscog-eem.1578 [not merging with Clément's immutability code just yet]
x64 Cogit:
Fix basicNew: for 4-byte arrays such as Bitmap.  The primitive needs to round down num slots and compute odd bits in the 64-bit form.

Provide facilities to map selectotrs back to selector indices when unlinking sends on 64-bit Spur (which has 32-bit inline caches).

Add method-relative labelling of disassembly if relativeAddressDisassembly option is asserted, to allow quick comparison of pre-compaction and post-compaction code (of course there was nothing wrong with compaction; the problem was with basicNew: but maybe it'll be useful one day ;-)).

Add a cr when printing just registers every instruction.