CogVm source as per VMMaker.oscog-eem.1630
Cogit:
Fix the bug in ceCaptureStackPointers caused by pushing VarBasReg to save it
offsetting the stack pointer.

Eliminate all the maybe*VarBase methods since we have hasVarBaseRegister and
the few clients can test that and hnce not have to divine if maybeSaveVarBase
pushed something or not.

On Mac IS X modify the stack alignment definitions to reflect documented
Apple policy.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3553 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 3c333af409fc69de79f6e691397e328b2cebd922