In this version, the Stack VM compiled to C is fully working with Immutability ON.
Load the package to try to use it.

I fixed the primitive not popping enough the stack...

Fixed a zip in Cogit.

Important change in the stack interpreter: if the cannotAssign:withValue: callback was called, its sender pc was ahead by one due to the fetchNextBytecode before the normalSend. I postponed the fetchNextBytecode after the immutability check, which works fine in the compiled VM. 

I think the Simulator works fine too but I might need to fix something there.