CogVM source as per VMMaker.oscog-eem.1656
Slang: Remember to remove the declaration for a local that is introduced for
inlining where the inlining uses the parse tree, not the local.  Otherwise,
stale declarations can be left that will mistype subsequent inlines of a method
with formals of the same name.  This fixes quite a few assignment type warnings.

Changes to interpreter bytecode dispatch (fetching next bytecode later)  to support optional IMMUATBILITY

primitiveStringReplace needs to check for immutability amd soon should support
16- and 64-bit access.  Provide an error code for immediate replacement.
primitiveStringReplace.

Add immutability check to changeClassOf:to: for primitiveAdoptInstance et al.

Cogit: Merge in ARM 16 bit write handling.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3563 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: c103d0368c954196514557a4c8b666c3fb6219ec