VMMaker.oscog-eem.1656 ARM Cogit: Fix slip in new 16-bit aware dispatchCOncretize.
Interpreter: Provide an error code for immediate replacement

Slang: Remember to remove the declaration for a local that is introduced for inlining where the inlining uses the parse tree, not the local.  Otherwise, stale declarations can be left that will mistype subsequent inlines of a method with formals of the same name.

Use declatationAt:put:, not direct access to declarations, for debugging.

Don't transform the parse tree/collect new locals etc, when inlining withouty renaming, just to cut down on the garbage.