VMMaker.oscog-eem.1661.mcz Cogit IMMUTABILITY cleanups. Remove the need to map the primitive error code store bytecode by having mapFor:bcpc:performUntil:arg: skip the CallPrimitive & StoreTemp bytecoees if present.  Also avoid the set-up overhead when mapping the first bytecode by moving the first invocation early.
Correct and simplify testBcToMcPcMappingForCogMethod:; is compiled method parameter was never used, and the startpc of a method should be derived from the cogmethod's method obj.

Refactor genImmutableCheck:slotIndex:sourceReg:scratchReg:needRestoreRcvr: to lose its unused popBoolean: keyword.

Implement IMMTABILITY in SimpleStackBasedCogit.  All combinations of SimpleStackBasedCogit x StackToRegisterMappingCogit x IMMTABILITY=true/false work on the reader image in the simulator.  THANKS Clément!!
Add SimpleStackBasedCogit>>putSelfInReceiverResultReg and use it where appropriate.

Fix a comment or three.

Oh we will, we will not be moved.
Oh we have im-mut-a-bil-i-ty
Just by a bit in every object header,
we will not be moved.

[ed: that's pinning, surely]