CogVM source as per VMMaker.oscog-eem.1662
Slang: Fix bad regression in type inference from VMMaker.oscog-eem.1587.
returnTypeForSend:in: can only default return types to sqInt for unknown
selectors.  The rewrite in VMMaker.oscog-eem.1587 inadvertently defaulted the
return type of known methods whose return type was yet to be determined to
sqInt.  The fix is to only default to sqInt if there is no known method for the
given selector.  This fixes, for instance, the weird flipping of the type of
32-bit Spur's headerWhileForwardingOf:,which should always be #sqLong.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3565 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: faebe4192e48d301c14e30dcde5029639eea31cb