VMMaker.oscog-eem.1676 NewspeakVM: Don't use cppIf: #NewspeakVM in VM code.  It is appropriate only in plugin code.  VM code should use the NewspeakVM variable from VMBasicConstants.
Slang:
Fix nilOrBooleanConditionFor: to also work for symbol arguments (such as #Newspeak).  Emit a warning if a symbol is used in VM code when a variable also exists and the symbol/variable is not one that is defined at compile time.

Provide an abstraction for namesDefinedAtCompileTime to avoid the #includes: call that existed in all clients.