VMMaker.oscog-eem.1693.mcz Fix a bug in printHex: for 64-bits (faulty arithmetic). Fix the remaining 64-bit issues in the ThreadedFFIPlugin.  The X64SysVFFIPlugin now passes the test suite.
The issues addressed are:
Masks for restricting and signing integral return values must be declared as at least long (again the fact that a manifest integer constat's type defaults to int causes problems).  Cast 1 << foo to 1 asUnsignedLong << foo in one place where it is required.

Use isFloatObject: rather than floatClass to test for floats since this works for both SmallFloat64 and BoxedFloat64 on 64-bit Spur.

Several pointer declarations needed to be e.g. <var: #ptr type: #'long *'>, not <var: #ptr type: #'int *'>.

Several casts to int of of firstIndexableField: (which has tyype void *) were discarded.

ExternalAddresses and the like need to be created with BytesPerWord or (self sizeof: #'void *'),, not 4.

Modify Slang to cast a manifest constant that is the receiver of a shift to long (via postpending 'L' to the numeric string).

Speed up ffiIntegerValueOf: by adding a tag test (harmless in our current crop of VMs) and cppIf: SPURVM as appropriate.

(N.B. not merging with VMMaker.oscog-EstebanLorenzano.1692.  The change there-in is incorrect.  Esteban and I have discussed and resolved the issue off line).