CogVM source as per VMMaker.oscog-eem.1699
First, it was Stephane Rollandin who found the perform: bug.  Apologies for the
misattribution.

Allow a plugin to override moduleName in generating its export table, via
moduleExportsName.  Hence fix compilation of the SqueakFFIPrims plugin as an
internal plugin.

Also use a single variable to hold moduleExportsName in the generated source.

Make the SqueakFFIPrims plugin internal in the Mac pharo builds.
Add a target to the Mac makefiles to allow printing a plugins prerequisites.
This to try and debug why plugins are not remade unless FooPlugin.lib and/or
FooPlugin.bundle are deleted.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3631 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: d14b56b05d504f4bce8ebc1a2b6cb9bfdb478bd6