Sources as per VMMaker 4.15.3
For Interpreter>>primitiveResume: add check for isContext: required for Squeak trunk as of Kernel-eem.1000 and above.

Background - see Kernel-eem.1000 in Squeak trunk, update comment repeated here for reference:
Make Process>>resume primitive.  Andreas fixed the ancestor of the Cog VM so that the resume primitive fails if the suspendedContext is not a context.  This renders Tim's suspendedCOntext ifNil: [^self primitiveFailed] guard obsolete.  Hence nuke primitiveResume.

Remove <inline: false> from Interpreter>>resume: because it is no longer required. The inliner is much improved since early days, and many other <inline: false> directives can probably also be removed, but that can be the subject of a later update.


git-svn-id: http://squeakvm.org/svn/squeak/trunk@3633 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 2d24f1ff8b678168c1e280b431c69e453600fa90