Fix immutability suport in the Cogit by making sure that the Slang output for the bytecode generatorTable marks inst var store bytecode generators as mapped if IMMUTABILITY is set.
Change vm parameter 65 to be a set of flags, not merely the boolean for MULTIPLE_BYTECODE_SETS, and make the flags answer MULTIPLE_BYTECODE_SETS & IMMUTABILITY.

Add a PC map printer to the Cogit (used to debug the above issue).  Make the current image facade support more of the printing api so that the map printer can be tested in-image.

Nuke an obsolete method.