Cogit: Revise the register saving convention for trampoline calls.  Instead of passing in a boolean to indicate if all registers should be saved or not (an aproach Ryan correctly notes is bogus), pass in a register mask of registers to be saved.  If the trampoline call has a machine code continuation then the argument is usually callerSavedRegMask, and on MIPS and ARM this is empty.
Use the above to move the saving of registers in the store-check into the store-check trampoline and out of the call of the trampoline.  Further, use genJumpImmediate: in the immediate test, which saves an instruction.

Nuke some obsolete methods.

Simulator:
Change the ByteCountsPerMicrosecond to better accord with a contemporary Mac.