CogVM source as per VMMaker.oscog-eem.1709
General:
Nuke unused named prim primitiveOriginalMillisecondClock so VM can eliminate
ioOldMSecs.

Nuke ioOldMSecs and ioLowResMSecs to simplify the tiume management and avoid
mistakes were we use different versions of clocks for different parts of the
same basic io operations.  Nuke SqueakUIFlushUseHighPercisionClock preference.

Change vm parameter 65 to be a set of flags, not merely the boolean for
MULTIPLE_BYTECODE_SETS, and make the flags answer MULTIPLE_BYTECODE_SETS &
IMMUTABILITY.

Cogit:
Revise the register saving convention for trampoline calls.  Instead of passing
in a boolean to indicate if all registers should be saved or not (an aproach
Ryan correctly notes is bogus), pass in a register mask of registers to be
saved.  If the trampoline call has a machine code continuation then the
argument is usually callerSavedRegMask, and on MIPS and ARM this is empty.

Use the above to move the saving of registers in the store-check into the
store-check trampoline and out of the call of the trampoline.  Further, use
genJumpImmediate: in the immediate test, which saves an instruction.

Fix immutability support in the Cogit by making sure that the Slang output for
the bytecode generatorTable marks inst var store bytecode generators as mapped
if IMMUTABILITY is set.

Add a PC map printer to the Cogit (used to debug the above issue).

Plugins:
Update primitiveFileRead for Spur.  It doesn't need the PrimErrObjectMayMove
handling since Spur has pinning.

Build system:
Fix dependency files so that they apply to all builds, production, assert &
debug, not just the one during which they were generated.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3635 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 3484735b1663a5e3ebb9e280be0bb33e41491dd2