VMMaker.oscog-nice.1735.mcz Workaround spurious breakage caused by nice.1732 Slang inliner does inline this call:
    sqInt shifted = 0xFFFFFFFFU;
    shifted = positive32BitIntegerFor(shifted);

into:

    sqInt shifted = 0xFFFFFFFFU;
    if( shifted <= MaxSmallInteger)
    {
        shifted = (shifted << 1) | 1;
       ...snip...

shifted interpreted as signed integer is negative.
MaxSmallInteger is a signed integer constant, so the if test is TRUE.
and a negative SmallInteger is created (with value -1) instead of a LargePositiveInteger as it should have been.

The intention is to interpret shifted as an unsigned integer, but since we did not help SLANG with a hint, slang did produce a signed (sqInt) by default.