VMMaker.oscog-nice.1736 Fix another bug of 1732.nice commit which broke primitiveBitAnd and primitiveBitOr
The symptom is this one at least in stack spur 64bits:
(-1 bitAnd: -1) -> -1 "correct"
(-1 perform: 1bitAnd: with: -1) -> 16r1FFFFFFF 'huh?"

it's -1 asUnsignedInteger >> 3...

The first one is working Ok because it's using bytecodePrimBitAnd which is very simple and correct.

here is the mistake I introduced a #usqInt type hint in slang.
At the time I did it, it was to reduce the number of signed mismatch warnings which are a plague. Yes but:

integerArgument := self positiveMachineIntegerValueOf: integerArgument.

Ouch, the return value is supposed to be unsigned.
While the parameter has to be signed in order to have the isIntegerObject: bit tricks working once inlined by slang...
So reusing the variable name for two different things was not a good idea finally (it rarely is).

This doesn't happen in my own branch because I've carefully rewritten the shift logic to use exclusively well defined unsigned behavior, but my changes are too aggressive, and I'd like to introduce them gradually to give a chance of reviewing them.

------------------

While at it, use cCorceSimple:to: 'unsigned int' when fetching integerValueOf rather than asUnsignedInteger.
The reason is that the first will never do sign extension, while the second would in 64bits.
OK, the 64bits spur cannot take this path, because a 32bit int would be a SmallInteger, but I prefer to throw good basis, this will avoid having to ask myelf about the same problem again in a couple of years...

I would much prefer to have a fetchUnsigned32/64, but we'll see that another time.