CogVM source as per VMMaker.oscog-eem.1756
General:
Correct undue sign extension wile promoting 32 to 64 bits int in
fetchLong64:ofObject: This was causing trouble in non spur object memory (V3)
especially since I decided to generalize the usage of this message.

Spur:
Have WideString>>at: fail to answer an out-of-range character in the
interpreter primitive (as well as in the JIT).

Cogit:
Fix the ARM's caller-saved register mask now that we can name all the actual
registers.

Spur Cogit:
Reworked machine code generation of immutability so for common stores it
uses a single trampoline for both store checks and immutability checks.

Add a compile-time flag, CheckRememberedInTrampoline, that controls whether the
remembered check is generated in-line or in the trampoline.  If checked in the
trampoline and on 64-bits use one memory access to get the header and share the
access between the IMMUTABILITY check and the rememberedBit check.

BitBltPlugin:
BitBltSimulation is operating on bits of 32bits words, therefore it's better
to declare its operands as 'unsigned int' rather than sqInt.

On 32bits VM, this should not change anything, but on 64bits spur, it makes
this snippet work:

    | wideString source pos blt expectedWideString |
    source := #[1 64 255 14 1 64 48 251].
    expectedWideString := WideString fromByteArray: source.
    wideString := WideString new: source size // 4.
    pos := 0.
    blt := (BitBlt
        toForm: (Form new hackBits: wideString))
        sourceForm: (Form new hackBits: source).
    blt
        combinationRule: Form over;
        sourceX: 0;
        sourceY: pos // 4;
        height: wideString byteSize // 4;
        width: 4;
        destX: 0;
        destY: 0;
        copyBits.
wideString restoreEndianness.
self assert: wideString = expectedWideString

Hence it fixes loading/diffing MCZ with wide character.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3666 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: b44b7c4dd41e78a9c093372901a6f1a8604a86a9