VMMaker.oscog-eem.1778.mcz Split the FullBlockClosure>>value[:...] support into its own set of primitives (531 through 538).  Make these optional, depending on SistaV1BytecodeSet. No longer follow the method and context fields in a closure in activateNewClosureMethod:numArgs:mayContextSwitch:; the caller will have failed if these are forwarders, so there is no need to check again.
Simulator:
Fix StackInterpreterSimulator>>endPCOf: for CompiledBlocks, where a blockReturn is the last bytecode of a block method.

Slang:
To allow methods to be optional upon a bytecode set add the name of the bytecode set as a key to the value true in initializationOptions in all of the bytecode set initializers, with names SqueakV3PlusClosuresBytecodeSet, NewsqueakV4BytecodeSet, SistaV1BytecodeSet.