VMMaker.oscog-eem.1779 Fix simulation of the LargeIntegersPlugin.  The issue is that interpreterProxy firstIndexableField: answers a CArray whose unitSize reflects the object's unit size, and becuase LargeIntegers are byte objects, the CArray is a byte accssor by default.  So add InterpreterPlugin>>firstIndexableField:as:, which uses cCoerce:to: to provide a CArray typed as requested, and use it everywhere the LargeIntegersPlugin takes the firstIndexableField of a LargeInteger, and cast to #'unsigned int *'.
Go some way to speed up the warppers by
a) improving CArray>at:[put:] to assume unitSize is a suitable power of two, and
b) add CFloatArray to access float or double

Later on we can add CUnsignedIntegerArray, as required.


N.B.  While this fixes simulation of the LargeIntegerPlugin there is still something wrong under simulation in whatever produces large fonts.  Start up the Squeak 5.0 release image with the "Squeak 5.0 Release Notes" text in huge letters in a workspace towards the front and you'll notice that in the simulator none of the large fonts are displayed, although the text from "Fast Become" works fine.