Fix simulation of coerceTo:sim: - a CArrayAccessor behaves like a CArray (with 0-based index) but translates indices to 1-based for accessing an underlying (fake) Smalltalk object.
In case of CArray, we can directly access the interpreter memory so we don't need a CArrayAccessor anyway.

The only case where we use asCArrayAccessor is that of BallonArray where we simulate both a Smalltalk fake simArray and the interpreter memory.  We absolutely need to invoke BallonArray>>at:[put:] to be sure that both are maintained, and that's what the CArrayAccessor indirection does. #asSingleFLoatAccessor would bypass this mechanism and is thus not appropriate. Please notice the += -1 offset in #asCArrayAccessor in this case.