VMMaker.oscog-nice.1798 1) If CogMethodZone is not given a chance to initializationOptions:,  then
theCCodeGenerator shouldIncludeMethodFor: CogMethodZone selector: #followForwardedLiteralsInOpenPICList

will answer false even for a SpurObjectMemory.

Let's give this chance to any ancillaryClasses: now that this solution has been blessed by Eliot mail :)

This restores the ability to generate suprsrc & co

2) Don't let LargeIntegersPlugin be the sole sender of stAt:put: -- this let a chance to remove some not so used VMMaker code and shrink the code area if we want to.

3) Factor out the duplicated code that is generating C integer literals
  The intention is to later change the generation, because it unecessarily generate unsigned long long when signed long long would do, and because it is not correct for long long negative int (the C Compiler can cope with this but would generate warnings) and even less correct for INT_MIN and LONG_LONG_MIN (which would be transformed as unsigned by C compiler and would most rarely meet our expectations).

4) Dispatch on the translated parse nodes for infering translated C type.
  By now, provide an exact image of previous behavior.

Type inference is for assigning a type to variables, but also essential for correctly inlining translated methods.

The mid term intention is to improve this type inference and use it in TMethod>>addTypesFor:to:in: too (which does quite the same type of job).

This may later help to relax the TMethod>>isFunctional which currently drastically limits the inlining capabilities.

These future changes may have uncontrolled side effects on generated code. They will be committed when presumed harmless for code generation of head revision (asserted thru lack of regression when running Squeak trunk tests), and in all cases after the burst of Sista activity,