CogVM source as per VMMaker.oscog-eem.1824
Cogit:
Add a primitive that answers pc map data for methods which can be
used to better decorate methods in the VM Profiler.  Refactor the pc
map enumeration facilities so that the Sista pic data primitive can
share the same enumerator.  Do this by collapsing the
isBackwardBranch and annotation parameters into a single parameter.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3674 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: ee2b8ac5eaddd24cd92ad4854367652fe95903d7