Unify type inference used to #inferReturnTypeFromReturnsIn: with that based on AST.
That means many more type inference for message sends (returnTypeForSend:in:). Ideally any message generated specially thru initializeCTranslationDictionary should have a corresponding entry in returnTypeForSend:in:, because type inference cannot apply to special tricks applied by CCodeGenerator.

There is still possible ambiguity related to different types in different path (if else/ ()?: ). This was the whole point of #addTypesFor:to:in:. IMO we cannot resolve automatically and maybe the best will be to raise a Warning, optionnally ignore it and let expert user manually fix it..

Now that type inference is a bit more thorough, release type restrictions in #isFunctional in order to obtain much more aggressive inlining.

With those, a gain of 10% is possible on the LargeIntegersPlugin dominated bench:
[ArbitraryPrecisionFloatTest suite run] timeToRun.