Change overflow tests in precondition rather than postcondition for small integer multiplication.
Because integer overflow is undefined behaviour, and because we can't rely on undefined behaviour, the compiler has a license to eliminate the post-condition as dead code (at least since iso c89 acception).
Post-condition currently works because we use -fwrapv compiler flag, but relying on specific compiler flags is bad practice in the long term. 

The pre-condition test is not much more expensive and we can avoid a isIntegerObject: test in post-condition, so there's no reason to not follow the C99 standard.

Anyway, Cog will detect and accelerate these multiplication with just (integer multiply, jump on overflow) instructions, so we're speaking of the fallback code here (like in case of byte code simulation).

Note that I used a 64 bits (sqLong) result and signedInteger64For: in my own branch to completely eliminate overflow in 32bits and go beyond SmallInteger limits. I did not reproduce it here because it's a departure from SmallInteger primitive that states:

	"Primitive. Multiply the receiver by the argument and answer with the
	result if it is a SmallInteger. Fail if the argument or the result is not a
	SmallInteger. Essential. No Lookup. See Object documentation whatIsAPrimitive."

Maybe we could implement this large integer trick in bytecodePrimMultiply (since it already has special case for Float argument, why not for large integer result...)...