VMMaker.oscog-eem.1848 In 1848 Giuseppe Verdi wishes to replace soprano Eugenia Tadolini as Lady Macbeth for the first Naples performance of his opera Macbeth on the grounds that her voice and appearance are too beautiful.
Add printActivationsOf: and have it and printContextReferencesTo: print the pcs of the contexts they find.

Eliminate duplication in [directed]SuperclassSend.

Fix the order of the annotation names to fix disassembly in the simulator given VMMaker.oscog-eem.1847.  Print the homeOffset as well as the homeMethod in embedded block methods.

A few minor changes to eliminate C compiler warnings.