VMMaker.oscog-eem.1850 In 1850, Harriet Tubman becomes an official conductor of the Underground Railroad in the United States. The Potato Famine is ongoing.
Partially fix inlining of returning ifs.  inlineCodeOrNilForStatement:in: is wrong to assume that a send is never a direct return; it is if it is in a returning if.  So refactor it to inlineCodeOrNilForStatement:returningNodes:in: and have its caller tryToInlineMethodsIn: compute the set of returning nodes appropriately.  This could all be cleaned up somewhat but better to get it working first.  Ficxing this bug then uncovers another in exitVar:label:. exitVar:label: uses replaceNodesIn: to replace ^exprs with exitVar := expr. goto exitLabel, but replaceNodesIn: is strictly top-down, so any replacement for ^expr ifTrue: [...^fu...] ifFalse: [...^bar...]
			 will prevent replacement of either ^fu or ^bar. The corollary is that ^expr ifTrue: [foo] ifFalse: [^bar]
			 must be transformed into expr ifTrue: [^foo] ifFalse: [^bar].


This isn't finished yet; e.g.

	SpurMemoryManager>>firstFixedFieldOfMaybeImmediate: oop
		"for the message send breakpoint; selectors can be immediates."
		<inline: false>
		^(self isImmediate: oop)
			ifTrue: [oop asVoidPointer]
			ifFalse: [self firstFixedField: oop]
	
is mistranslated as

	static void * NoDbgRegParms
	firstFixedFieldOfMaybeImmediate(sqInt oop)
	{
	    return (oop & (tagMask())
	        ? ((void *)oop)
	        : (/* begin firstFixedField: */
	            return pointerForOop(oop + BaseHeaderSize)));
	}

but it's close enough, given that the generated code from the previous version is wrong yet compiles silently.  This versiomn correctly translates e.g. inlines of positiveMachineIntegerFor: and inlines of positive32BitIntegerFor: into it.  Hopefully we'll fix the remaining issues PDQ.

Change statement list printing to use KB style I can't read anything else, forgive me).