VMMaker.oscog-eem.1851.mcz Slang: Fix inlining in the wake of recent type inferrence/promotion changes that changed the inlining landscape sufficiently to uncover bugs.
inlineCodeOrNilForStatement:returningNodes:in: and inlineSend:directReturn:exitVar:in: must cooperate to not drop returns in statement lists, other than at the end of the statement list, in the context of returning ifs.  They must be careful not to leave returns at the end of statement lists in returning ifs.  The symptom was falure to assign the value of a converted SmallInteger when inlining positive32BitIntegerFor: into positiveMachineIntegerFor: in the 32-bit case.

Do not use cppIf: SPURVM ifTrue: ... or cppIf: MULTIPLEBYTECODESETS ifTrue: in InterpreterPrimitives and subclasses, and allow the vmClass to override VMBasicConstants so as to eliminate code at translation time that tests COGVM, COGMTVM, SPURVM  and STACKVM.  These need to be defined at compile time only in plugins; the VMs generated are soecific to the values of these constants.

Rewrite methodClassOf: to not reply on comma-separated expressions in the generated C.

In 1851 a population census is taken in the United Kingdom. The population reaches 21 million. 6.3 million live in cities of 20,000 or more in England and Wales and cities of 20,000 or more account for 35% of the total English population.  The first protected submarine telegraph cable is laid, across the English Channel. The Chilean Revolution ends after the rebel defeat at the Battle of Loncomilla.