CogVM source as per VMMaker.oscog-eem.1851
Slang:
Fix inlining in the wake of recent type inferrence/promotion changes that
changed the inlining landscape sufficiently to uncover bugs.

inlineCodeOrNilForStatement:returningNodes:in: and inlineSend:directReturn:exitVar:in:
must cooperate to not drop returns in statement lists, other than at the end of
the statement list, in the context of returning ifs.  They must be careful not
to leave returns at the end of statement lists in returning ifs.  A symptom
was falure to assign the value of a converted SmallInteger when inlining
positive32BitIntegerFor: into positiveMachineIntegerFor: in the 32-bit case.

Do not use cppIf: SPURVM ifTrue: ... or cppIf: MULTIPLEBYTECODESETS ifTrue: in
InterpreterPrimitives and subclasses, and allow the vmClass to override
VMBasicConstants so as to eliminate code at translation time that tests COGVM,
COGMTVM, SPURVM  and STACKVM.  These need to be defined at compile time only
in plugins; the VMs generated are soecific to the values of these constants.

Rewrite methodClassOf: to not reply on comma-separated expressions in the
generated C.

Regenerate plugins using the recently revised type inferrence code.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3689 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 057a6351bed0f434a5510d525cdbf834d04fb973