CogVM source as per VMMaker.oscog-nice.1853
Fix missing comparison operation in genPrimitiveIdentical on V3. This should
fix Character = problem encounterd in Cuis.

Comment concretizePrefetchAw, and store machineCode in last statement so as to
avoid a -Wunused-value warning.

Correct slip in primitiveMethodXray, the flags were not set in each caseOf:
branch (thanks to -Wunused-value).

Remove assert in classTagForClass: because ensureBehaviorHash: already does it.



git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3690 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 6704f67a7c404cf35c32e376ce63954d861d8adf