This commit breaks the CMake generation, but there is a lot of other work  I want commited.
The addition of the various 'source directories' in addition to the old 'src' directory that originally existed for Cog ..i.e. spursrc, stacksrc...etc has introduced some complexity that has to be accounted for.

I will think about this next, I just want to get the documentation work stored for now.