VMMaker.oscog-eem.1861 Fix the Spur implementation of becomeForward:copyHash: for becomming immediates with copyHash false.  Fix the scavenger's referenceCountRememberedReferents: for forwasrding to immediates, and add a printRememberedSet convenience.
Make primitiveDoPrimitiveWithArgs usable as a mirror primitive, such that no message needs to be sent to the receiver to evaluate a primitive on it.  This would make possible Context>>receiver: anObject tryPrimitive: primIndex withArgs: argArray.

Use symbolic constants for the primitive numbers of the primitive-executing primitives.  Provide support for running the leak checker after an external primtiive call in the StackInterpreter (for debugging leaks in the GdbARMPlugin).