CogVM source as per VMMaker.oscog-eem.1865
Fix two bugs in ephemerons (found by Gille Polito - thanks!).

1. obj stack extension must happen with marking set to true if new pages aren't
   to be GC'ed.  So move the assignments to marking out into markObjects: where
   they can surround nilUnmarkedWeaklingSlots.

2. fireAllUnscannedEphemerons needs to enumerate the unscannedEphemerons set
   properly, an oop at a time.

Use bytesPerOop instead of wordSize throughout unscannedEphemerons access for
consistency.

Fix the Spur implementation of becomeForward:copyHash: for becomming immediates
with copyHash false.  Fix the scavenger's referenceCountRememberedReferents:
for forwarding to immediates, and add a printRememberedSet convenience.

Make primitiveDoPrimitiveWithArgs usable as a mirror primitive, such that no
message needs to be sent to the receiver to evaluate a primitive on it.  This
would make possible Context>>receiver: obj tryPrimitive: idx withArgs: args.

Use symbolic constants for the primitive numbers of the primitive-executing
primitives.  Provide support for running the leak checker after an external
primtiive call in the StackInterpreter (for debugging leaks in t' GdbARMPlugin).


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3728 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 93f9c0bb4298ff4551ced98074bc9b26c3453ba4