Fix slang regression in VMMaker.oscog-eem.1868.  See MessageNode>>asTranslatorNodeIn:.
Fix ephemerons (at least for Gule's test case).  The mournQueue is of course a root for teh scavenger and so must be visited at the start of a scavenge.  Fix receiver in fireEphemeron:.

Simulator:  Make Spur clone the heap and do a lemming scavenge if leakCheckNewSpaceGC is set, just as ObjectMemory does.  Update Spur's leak checker to scan the mournQueue.