Anti-slavery society forms (NY)...
When I benched Immutability/write barrier, games benchmarks had identical performance with and without it (the difference was within noise).

However I built a microbench doing only inst var stores:

Foo>>imm: imm nonImm: nonImm
	iv1 := 1.
	iv2 := #foo. 
	iv3 := imm.
	iv4 := nonImm.

| f |
f := Foo new.
[f imm: 2 nonImm: #bar ] bench   
	
And in this case there was 18.7% overhead with the write barrier. BinaryTree bench has a similar method, use it extensively and yet we can't see the overhead on the whole bench, so it's not that critical, but it's still something which could matter in specific cases.

In this commit I changed the JIT so that (with Immutability/Write barrier ON) methods with only frameless instance variable stores are compiled with two paths. The code first checks if the receiver is immutable/read-only, and takes the right path accordingly. The first path is frameless and does not include immutability/write barrier checks. The second one is frameful and does all the immutability/write barrier checks.

I didn't do it for blocks as I believe blocks with only instance variable store are not that common - I may be wrong, if you prove to me I am wrong on this with numbers, I will do it for blocks.

In the micro-bench shown, #imm:nonImm:, the overhead decreased from 18.7% to 2.6%.

I think this has an impact in general to setter methods.

I made that in a Pharo sprint (or what it matters).