CogVM source as per VMMaker.oscog-eem.1876
Cogit:
Change the JIT so that (with Immutability/Write barrier ON) methods with only
frameless instance variable stores are compiled with two paths. The code first
checks if the receiver is immutable/read-only, and takes the right path
accordingly. The first path is frameless and does not include immutability/write
barrier checks. The second one is frameful and does all the immutability/write
barrier checks.

Eliminate the callerSavedRegMask variable and replace it with a
CallerSavedRegisterMask variable computed at register initialization time.

On ARM:
- define the caller-saved registers as the subset of the caller-saved registers
  that excludes R0/CArg0Reg/TempReg & R1/CArg1Reg (R0 is the C result register),
  and use R2 & R3 for abstract registers (ClassReg and Arg0Reg).
- save and restore elements of the caller-saved registers as appropriate around
  run-time calls.
- use LDM & STM to do the saving.
- hence provide two extra registers, now provinding Extra[012]Reg

Fix some compilation warnings in cogitX64.c

Plugins:
Change the JPEGReaderPlugin to handle 8- and 16- bit JPEGs (thank you Laura ! ;
still to integrate the refactoring into C and Smalltalk code).

More inlining of low-level copy routines in the LargeInegersPlugin.


git-svn-id: http://squeakvm.org/svn/squeak/branches/Cog@3733 fa1542d4-bde8-0310-ad64-8ed1123d492a

Former-commit-id: 5cdcaf6a0ce9eaa2ea8356b80074168727496c31